requirements.txt
streamlit
pandas
numpy
statsmodels
scikit-learn
plotly
sqlalchemy
python-dotenv
pymysql
